db-model: dataset -> crucible_dataset
