+++ nexus/db-model/src/lib.rs
@@ -21,7 +21,7 @@ mod clickhouse_policy;
 mod cockroachdb_node_id;
 mod collection;
 mod console_session;
-mod dataset;
+mod crucible_dataset;
 mod dataset_kind;
 mod db_metadata;
 mod device_auth;
@@ -141,7 +141,7 @@ pub use clickhouse_policy::*;
 pub use cockroachdb_node_id::*;
 pub use collection::*;
 pub use console_session::*;
-pub use dataset::*;
+pub use crucible_dataset::*;
 pub use dataset_kind::*;
 pub use db_metadata::*;
 pub use deployment::*;
