+++ schema/crdb/dbinit.sql
@@ -557,9 +557,10 @@ CREATE TYPE IF NOT EXISTS omicron.public.dataset_kind AS ENUM (
 );
 
 /*
- * A dataset of allocated space within a zpool.
+ * Table tracking the contact information and size used by datasets associated
+ * with Crucible zones.
  */
-CREATE TABLE IF NOT EXISTS omicron.public.dataset (
+CREATE TABLE IF NOT EXISTS omicron.public.crucible_dataset (
     /* Identity metadata (asset) */
     id UUID PRIMARY KEY,
     time_created TIMESTAMPTZ NOT NULL,
@@ -570,56 +571,29 @@ CREATE TABLE IF NOT EXISTS omicron.public.dataset (
     /* FK into the Pool table */
     pool_id UUID NOT NULL,
 
-    /* Contact information for the dataset */
-    ip INET,
-    port INT4 CHECK (port BETWEEN 0 AND 65535),
-
-    kind omicron.public.dataset_kind NOT NULL,
+    /*
+     * Contact information for the dataset: socket address of the Crucible
+     * downstairs service that owns this dataset
+     */
+    ip INET NOT NULL,
+    port INT4 CHECK (port BETWEEN 0 AND 65535) NOT NULL,
 
     /* An upper bound on the amount of space that might be in-use */
-    size_used INT,
-
-    /* Only valid if kind = zone -- the name of this zone */
-    zone_name TEXT,
-
-    quota INT8,
-    reservation INT8,
-    compression TEXT,
-
-    /* Crucible must make use of 'size_used'; other datasets manage their own storage */
-    CONSTRAINT size_used_column_set_for_crucible CHECK (
-      (kind != 'crucible') OR
-      (kind = 'crucible' AND size_used IS NOT NULL)
-    ),
-
-    CONSTRAINT ip_and_port_set_for_crucible CHECK (
-      (kind != 'crucible') OR
-      (kind = 'crucible' AND ip IS NOT NULL and port IS NOT NULL)
-    ),
-
-    CONSTRAINT zone_name_for_zone_kind CHECK (
-      (kind != 'zone') OR
-      (kind = 'zone' AND zone_name IS NOT NULL)
-    )
+    size_used INT NOT NULL,
 );
 
-/* Create an index on the size usage for Crucible's allocation */
-CREATE INDEX IF NOT EXISTS lookup_dataset_by_size_used_crucible on omicron.public.dataset (
-    size_used
-) WHERE size_used IS NOT NULL AND time_deleted IS NULL AND kind = 'crucible';
-
-/* Create an index on the size usage for any dataset */
-CREATE INDEX IF NOT EXISTS lookup_dataset_by_size_used on omicron.public.dataset (
-    size_used
-) WHERE size_used IS NOT NULL AND time_deleted IS NULL;
+/* Create an index on the size usage for any Crucible dataset */
+CREATE INDEX IF NOT EXISTS lookup_crucible_dataset_by_size_used ON
+    omicron.public.crucible_dataset (size_used)
+  WHERE time_deleted IS NULL;
 
 /* Create an index on the zpool id */
-CREATE INDEX IF NOT EXISTS lookup_dataset_by_zpool on omicron.public.dataset (
-    pool_id,
-    id
-) WHERE pool_id IS NOT NULL AND time_deleted IS NULL;
+CREATE INDEX IF NOT EXISTS lookup_crucible_dataset_by_zpool ON
+    omicron.public.crucible_dataset (pool_id, id)
+  WHERE time_deleted IS NULL;
 
-CREATE INDEX IF NOT EXISTS lookup_dataset_by_ip on omicron.public.dataset (ip);
+CREATE INDEX IF NOT EXISTS lookup_crucible_dataset_by_ip ON
+  omicron.public.crucible_dataset (ip);
 
 /*
  * A region of space allocated to Crucible Downstairs, within a dataset.
